requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.29
aiomysql==0.2.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
qdrant-client==1.7.4
requests==2.31.0
trafilatura==1.6.3
python-docx==1.1.0
pypdf==4.1.0
jinja2==3.1.3
python-multipart==0.0.9
pytest==8.1.1
httpx==0.27.0
itsdangerous==2.1.2
orjson==3.10.0
